+++ frontend/src/components/index.js
@@ -1,13 +1,15 @@
-import PrivateRoute from "./PrivateRoute";
-import LoadingIndicator from "./LoadingIndicator";
 import BackgroundWrapper from "./BackgroundWrapper";
-import ParticleWallpaper from "./ParticleWallpaper";
 import CampaignCard from "./CampaignCard";
+import LoadingIndicator from "./LoadingIndicator";
+import ParticleWallpaper from "./ParticleWallpaper";
+import PrivateRoute from "./PrivateRoute";
+import ReviewerStepper from "./ReviewerStepper";
 
 export {
-  PrivateRoute,
-  LoadingIndicator,
   BackgroundWrapper,
-  ParticleWallpaper,
   CampaignCard,
+  LoadingIndicator,
+  ParticleWallpaper,
+  PrivateRoute,
+  ReviewerStepper,
 };
