+++ frontend/src/pages/rankings/FinalRatingApplicationComments/index.js
