+++ frontend/src/pages/rankings/FinalRatingCandidateCard/finalRatingCandidateCard.styled.js
