+++ frontend/src/pages/rankings/FinalRatingCandidateCard/index.js
